feat: Frontend - Split Class Roster's components.
