+++ src/components/DashboardLaunchButton/index.scss
@@ -1,0 +1,20 @@
+.dashboard-launch-container {
+    margin-left: 10px;
+    margin-top: 20px;
+  }
+
+  .main-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 30px;
+  }
+
+  .instructor-dashboard-button {
+    margin-left: auto;
+    margin-right: 10px;
+  }
+
+  .error-container {
+    margin-top: 20px;
+  }
