+++ src/constants/index.js
@@ -1,1 +1,2 @@
 export const SKILLABLE_URL = `${process.env.LMS_BASE_URL}/skillable_plugin`;
+export const defaultErrorMessage = 'An unexpected error occurred. Please try again later.';
